+++ doc/source/changelog.rst
@@ -18,6 +18,9 @@ MontePy Changelog
 #Next Release#
 --------------
 
+**Features Added**
+
+* Make `LatticeType.RECTANGULAR` and `LatticeType.HEXAHEDRAL` synonymous (:issue:`808`).
 
 **Bugs Fixed**